+++ index.js
@@ -4,7 +4,7 @@ const fs = require('fs');
 const config = require('./config.json');
 const { token } = require('./token.json');
 
-const client = new Client({ intents: [Intents.FLAGS.GUILDS] });
+const client = new Client({ intents: [Intents.FLAGS.GUILDS, Intents.FLAGS.GUILD_PRESENCES, Intents.FLAGS.GUILD_MEMBERS, Intents.FLAGS.GUILD_MESSAGES] });
 
 const sequelizeRippedAssets = new Sequelize({
     dialect: 'sqlite',